# autopilot-api/src/modules/chatbot/dto/update-knowledge-document.dto.ts
import { IsOptional, IsString, IsUUID, MaxLength, MinLength } from 'class-validator';

export class UpdateKnowledgeDocumentDto {
  @IsUUID()
  tenantId: string;

  @IsOptional()
  @IsUUID()
  workspaceId?: string;

  @IsString()
  @MinLength(1)
  @MaxLength(255)
  title: string;
}